+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## 0.8.7 - 2025-06-15
 
 ### Added
 
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Add fieldset label on checkbox group and radio group components
 - Add toggle on form and input components
 - Add custom events on pagination click
+- Add `href` rest attribute on tab component
 
 ### Changed